requirements.txt
# Web Framework
flask==3.0.0
werkzeug==3.0.1
fastapi==0.115.0
uvicorn==0.30.0
gunicorn==22.0.0

# Vector Database & Embeddings
chromadb==0.4.24
sentence-transformers==3.0.1
qdrant-client==1.9.0

# Retrieval & Ranking
rank-bm25==0.2.2

# HTTP & Web Scraping
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.2

# LLM Integration
ollama==0.6.1

# Document Processing (PDF)
pdfplumber==0.10.3

# Configuration
python-dotenv==1.0.1

# Numerical Computing
numpy<2.0

# Database (PostgreSQL)
sqlalchemy==2.0.0
psycopg2-binary==2.9.9

# Auth
pydantic==2.7.0
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.3

# Async / Background Jobs
aiofiles==23.2.1
redis==5.0.0
rq==1.16.0
